pyproject.toml
[tool.poetry]
name = "realtime"
version = "0.1.0"
description = ""
authors = ["WrldEngine <kamscienceit123.kamran@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
pydub = "^0.25.1"
python-dotenv = "^1.0.1"
pyaudio = "^0.2.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
